Extract material parse logic to filament-matp
