+++ libs/filament-matp/src/DirIncluder.cpp
@@ -14,13 +14,13 @@
  * limitations under the License.
  */
 
-#include "DirIncluder.h"
+#include <filament-matp/DirIncluder.h>
 
 #include <utils/Log.h>
 
 #include <fstream>
 
-namespace matc {
+namespace matp {
 
 bool DirIncluder::operator()(const utils::CString& includedBy, filamat::IncludeResult& result) {
     auto getHeaderPath = [&result, &includedBy, this]() {
@@ -66,5 +66,5 @@ bool DirIncluder::operator()(const utils::CString& includedBy, filamat::IncludeR
     return true;
 }
 
-} // namespace matc
+} // namespace matp
