+++ libs/filament-matp/src/JsonishLexer.cpp
@@ -14,11 +14,11 @@
  * limitations under the License.
  */
 
-#include "JsonishLexer.h"
+#include <filament-matp/JsonishLexer.h>
 
 #include <string.h>
 
-namespace matc {
+namespace matp {
 
 JsonType JsonishLexer::readIdentifier() noexcept {
     const char* lexemeStart = mCursor;
@@ -148,4 +148,4 @@ bool JsonishLexer::readLexeme() noexcept {
     return true;
 }
 
-} // namespace matc
+} // namespace matp
