+++ libs/filament-matp/src/MaterialLexer.cpp
@@ -14,9 +14,9 @@
  * limitations under the License.
  */
 
-#include "MaterialLexer.h"
+#include <filament-matp/MaterialLexer.h>
 
-namespace matc {
+namespace matp {
 
 bool MaterialLexer::readBlock() noexcept {
     size_t braceCount = 0;
@@ -97,4 +97,4 @@ bool MaterialLexer::readLexeme() noexcept {
     return nextMaterialType != UNKNOWN;
 }
 
-} // namespace matc
+} // namespace matp
