--- tools/matc/tests/Dir/Bar.h
@@ -1,1 +1,0 @@
-// Bar.h
