--- tools/matc/tests/Dir/Baz.h
@@ -1,1 +1,0 @@
-// Baz.h
