--- tools/matc/tests/Foo.h
@@ -1,1 +1,0 @@
-// test include file
